style: address review comments
