+++ v2/interface.go
@@ -1,8 +1,6 @@
 package v2
 
 import (
-	"context"
-
 	"github.com/confluentinc/confluent-kafka-go/kafka"
 
 	"github.com/honestbank/kp/v2/internal/middleware"
@@ -15,5 +13,5 @@ type KafkaProcessor[MessageType any] interface {
 	WithDeadletter(deadLetterTopic string) (KafkaProcessor[MessageType], error)
 	WithDeadletterOrPanic(deadletterTopic string) KafkaProcessor[MessageType]
 	Stop()
-	Run(processor func(ctx context.Context, message MessageType) error) error
+	Run(processor Processor[MessageType]) error
 }
